refactor inline palette
