--- components/common/CharmEditor/components/@bangle.io/js-lib/inline-palette/index.ts
@@ -1,6 +1,0 @@
-import * as inlinePalette from './inline-palette';
-
-export * from './hooks';
-export * from './inline-palette';
-export { inlinePalette };
-
