+++ components/common/CharmEditor/components/@bangle.io/lib/ui-components/index.ts
@@ -1,2 +1,1 @@
-export * from "./InlinePaletteRow/InlinePaletteRow";
 export * from "./UniversalPalette";
