+++ components/common/CharmEditor/components/inlinePalette/index.ts
@@ -1,0 +1,5 @@
+import InlineCommandPalette from './components/InlineCommandPalette';
+
+export * from './inlinePalette';
+
+export default InlineCommandPalette;
