+++ components/common/CharmEditor/components/nestedPage/NestedPagesList.tsx
@@ -1,13 +1,12 @@
 
 import { useEditorViewContext, usePluginState } from '@bangle.dev/react';
 import { useTheme } from '@emotion/react';
-import { Box, ClickAwayListener, MenuItem } from '@mui/material';
+import { Menu, MenuItem, ListItemIcon } from '@mui/material';
 import PageIcon from 'components/common/PageLayout/components/PageIcon';
 import PageTitle from 'components/common/PageLayout/components/PageTitle';
 import { usePages } from 'hooks/usePages';
 import { Page, PageContent } from 'models';
 import { useCallback } from 'react';
-import { createPortal } from 'react-dom';
 import { isTruthy } from 'lib/utilities/types';
 import useNestedPage from './hooks/useNestedPage';
 import { hideSuggestionsTooltip } from '../@bangle.dev/tooltip/suggest-tooltip';
@@ -32,48 +31,41 @@ export default function NestedPagesList () {
     [view]
   );
 
-  if (isVisible) {
-    return createPortal(
-      <ClickAwayListener onClickAway={() => {
-        hideSuggestionsTooltip(NestedPagePluginKey)(view.state, view.dispatch, view);
-      }}
-      >
-        <Box>
-          {Object.values(pages).filter(isTruthy).map(page => {
-            const docContent = ((page.content) as PageContent)?.content;
-            const isEditorEmpty = docContent && (docContent.length <= 1
+  function onClose () {
+    hideSuggestionsTooltip(NestedPagePluginKey)(view.state, view.dispatch, view);
+  }
+
+  return (
+    <Menu anchorEl={tooltipContentDOM} open={isVisible} onClose={onClose} sx={{ maxHeight: 400 }}>
+      {Object.values(pages).filter(isTruthy).map(page => {
+        const docContent = ((page.content) as PageContent)?.content;
+        const isEditorEmpty = docContent && (docContent.length <= 1
               && (!docContent[0] || (docContent[0] as PageContent)?.content?.length === 0));
 
-            return (
-              <MenuItem
+        return (
+          <MenuItem
+            sx={{
+              background: theme.palette.background.light
+            }}
+            onClick={() => onSelectPage(page)}
+            key={page.id}
+          >
+            <>
+              <ListItemIcon>
+                <PageIcon icon={page.icon} isEditorEmpty={Boolean(isEditorEmpty)} pageType={page.type} />
+              </ListItemIcon>
+              <PageTitle
+                hasContent={page.title.length === 0}
                 sx={{
-                  background: theme.palette.background.light
+                  fontWeight: 'bold'
                 }}
-                onClick={() => onSelectPage(page)}
-                key={page.id}
               >
-                <>
-                  <div>
-                    <PageIcon icon={page.icon} isEditorEmpty={Boolean(isEditorEmpty)} pageType={page.type} />
-                  </div>
-                  <PageTitle
-                    hasContent={page.title.length === 0}
-                    sx={{
-                      fontWeight: 'bold',
-                      height: 26,
-                      lineHeight: 26
-                    }}
-                  >
-                    {page.title.length !== 0 ? page.title : 'Untitled'}
-                  </PageTitle>
-                </>
-              </MenuItem>
-            );
-          })}
-        </Box>
-      </ClickAwayListener>,
-      tooltipContentDOM
-    );
-  }
-  return null;
+                {page.title.length !== 0 ? page.title : 'Untitled'}
+              </PageTitle>
+            </>
+          </MenuItem>
+        );
+      })}
+    </Menu>
+  );
 }
